Move template data to Jekyll collection
We're [moving the template
documentation](https://github.com/18F/federalist-uswds-template/pull/19) to
federalist-docs. By moving the template data to a Jekyll collection, they will
render as their own pages but also are usable as site data for the template
index page. We include a layout for the template pages so they're all consistent
in how they present the documentation.

This only adds documentation for the Basic Report template.